+++ R/zzz.R
@@ -1,5 +1,6 @@
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("This version of alterryx was built and tested
-                         for Alteryx Server 2018.4. Support for earlier
-                         versions is not guaranteed.")
+  message <- paste("This version of alterryx was built and tested for",
+                   "Alteryx Server 2018.4. Support for earlier versions is",
+                   "not guaranteed")
+  packageStartupMessage(message)
 }